Add confirm before deleting

+++ js/app.controller.js
@@ -69,6 +69,8 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    const usrConfirm = confirm('Are you sure you want to delete the location')
+    if(!usrConfirm) return
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
